requirements.txt
numpy>=1.19.0
pandas>=1.0.0
scikit-learn>=0.23.0
matplotlib>=3.3.0
seaborn>=0.11.0
jupyter>=1.0.0
biopython>=1.78
tqdm>=4.50.0
streamlit>=1.10.0
tensorflow-cpu>=2.6.0 